clean up includes, namespaces, and file separations
